script: move to separate directory

+++ scripts/shutdown-local-ubuntu.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -ex
 
+if ! [[ "$0" =~ "scripts/shutdown-local-ubuntu.sh" ]]; then
+    echo "must be run from repository root"
+    exit 255
+fi
+
 echo "killing etcdlabs server (sending SIGINT)"
 kill -2 $(lsof -ti tcp:2200)
